Add files via upload
Added Chatbot so customers can contact directly without the phone.  Made 2 one of the Bots, one with ai which is not deployed yet, just going with a basic html css javascript to test it out. more was done but not sure how to word it correctly lol. but good progress